// src/app/shared/model/pages.model.ts
export interface pageSelection {
  skip: number;
  limit: number
}
export interface Star {
  show?: boolean;
  half?: boolean;
}
export interface apiResultFormat {
  data: [];
  totalData: number;
}
// ...
export interface manageUsers {
  isSelected: boolean;
  id: number;
  customer_name: string;
  customer_image: string;
  customer_no: string;
  phone: string;
  email: string;
  location: string;
  created: string;
  last_activity: string;
  status: string;
  isStarActive: boolean;
}